+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 0.102.0 (Unreleased)
 
+FEATURES:
+
+* provider: allow creating `public` dns zones along with `private` and `public-private` once
+* provider: allow updating dns zone visibility (all transitions are allowed except `public-private -> public` one)
+
 ## 0.101.0 (October 31, 2023)
 
 ENHANCEMENTS: